Bring up to date w/ fork

+++ apikey_rotator.sh
@@ -1,11 +1,18 @@
-#!/bin/bash -e
-set -o pipefail
+#!/bin/bash
+set -eo pipefail
 
 APP_HOSTNAME=webapp1/tomcat_host
 
-                        # rotate API key
-                        # write new key to nondescript file in shared volume
-api_key=$(docker-compose exec -T cli conjur host rotate_api_key --host $APP_HOSTNAME)
-echo $api_key > local_foo
-docker cp local_foo $(docker-compose ps -q cli):/data/foo
-rm local_foo
+		# rotate api key for host
+api_key=$(docker-compose exec -T cli \
+			conjur host rotate_api_key --host $APP_HOSTNAME)
+
+		# if no arg provided
+if [[ "$1" == "" ]]; then
+		# write new key to nondescript file in shared volume
+	echo $api_key > local_foo
+	docker cp local_foo $(docker-compose ps -q cli):/data/foo
+	rm local_foo
+	sleep 5
+	docker-compose exec cli rm /data/foo
+fi
